Implemented raw 'country streamer', that counts amount of posts per country per hour.

+++ src/main/scala/com/eleks/socialnetworkanalyser/entities/CountryStats.scala
@@ -1,5 +1,5 @@
 package com.eleks.socialnetworkanalyser.entities
 
 case class CountryStats(country : String,
-                        postCount : Int
+                        postCount : Long
                        )
